+++ python/oneflow/framework/tensor.py
@@ -830,39 +830,35 @@ def _t(self):
     return flow._C.t(self)
 
 
-def _topk(self, k, dim, largest, sorted):
-    return self.topk(k, dim, largest, sorted)
+def _topk(self, k, dim: int = None, largest: bool = True, sorted: bool = True):
+    return flow.topk(self, k, dim, largest, sorted)
 
 
-def _nms(self, scores, iou_threshold: float):
-    score_inds = flow.argsort(scores, dim=0, descending=True)
-    boxes = flow._C.gather(self, score_inds, axis=0)
-    keep = flow._C.nms(boxes, iou_threshold)
-    index = flow.squeeze(flow.argwhere(keep), dim=[1])
-    return flow._C.gather(score_inds, index, axis=0)
+def _nms(boxes, scores, iou_threshold: float):
+    return flow.nms(boxes, scores, iou_threshold)
 
 
-def _nonzero(self, as_tuple):
+def _nonzero(self, as_tuple=False):
     return flow.nonzero(self, as_tuple)
 
 
-def _max(self, dim, keepdim):
+def _max(self, dim=None, keepdim=False):
     return flow.max(self, dim, keepdim)
 
 
-def _min(self, dim, keepdim):
+def _min(self, dim=None, keepdim=False):
     return flow.min(self, dim, keepdim)
 
 
-def _sum(self, dim, keepdim):
+def _sum(self, dim=None, keepdim=False):
     return flow.sum(self, dim, keepdim)
 
 
-def _mean(self, dim, keepdim):
+def _mean(self, dim=None, keepdim=False):
     return flow.mean(self, dim, keepdim)
 
 
-def _prod(self, dim, keepdim):
+def _prod(self, dim=None, keepdim=False):
     return flow.prod(self, dim, keepdim)
