+++ python/oneflow/nn/modules/nms.py
@@ -13,28 +13,14 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from typing import Optional
-
-import numpy as np
-
 import oneflow as flow
 from oneflow.framework.tensor import register_tensor_op
 from oneflow.nn.module import Module
 
 
-def nonzero_op(input, as_tuple=False):
-    if as_tuple and not input.ndim:
-        input = input.unsqueeze(0)
-    (res, size) = flow._C.argwhere(input)
-    slice_tup_list = [[0, int(size.numpy()), 1]]
-    res = flow.slice(res, slice_tup_list=slice_tup_list)
-    if as_tuple:
-        return tuple([flow._C.transpose(res, [1, 0])[x] for x in range(res.shape[1])])
-    else:
-        return res
-
-
-if __name__ == "__main__":
-    import doctest
-
-    doctest.testmod(raise_on_error=True)
+def nms_op(boxes, scores, iou_threshold: float):
+    score_inds = flow.argsort(scores, dim=0, descending=True)
+    boxes = flow._C.gather(boxes, score_inds, axis=0)
+    keep = flow._C.nms(boxes, iou_threshold)
+    index = flow.squeeze(flow.argwhere(keep), dim=[1])
+    return flow._C.gather(score_inds, index, axis=0)
