+++ python/oneflow/nn/modules/reduce_ops.py
@@ -23,35 +23,6 @@ from oneflow.nn.modules.utils import _check_axis
 
 
 def max_op(input, dim=None, keepdim=False):
-    """Computes the maximum value of all elements in the input tensor.
-    
-    Args:
-        input (oneflow.Tensor): the Input Tensor
-        dim (int, optional): the dimension to reduce. Default: `None`
-        keepdim (bool, optional): whether the output tensor has dim retained or not. Default: `False`
-
-    Returns:
-        Tensor or Tuple(oneflow.Tensor, oneflow.Tensor(dtype=int64)): If :attr:`dim` is `None`, returns 
-        the maximum value of all elements in the `input` tensor. Otherwise, returns a tuple of Tensor (values, indices), 
-        where the `values` are the maximum value of all elements in the `input` tensor,
-        the `indices` are the indices of the elements in the original input tensor.
-
-    For example:
-
-    .. code-block:: python
-
-        >>> import oneflow as flow
-        >>> input = flow.Tensor([[4, 1, 5], [2, 6, 3]])
-        >>> flow.max(input)
-        tensor(6., dtype=oneflow.float32)
-        >>> (values, indices) = flow.max(input, dim=1)
-        >>> values
-        tensor([5., 6.], dtype=oneflow.float32)
-        >>> indices
-        tensor([2, 1], dtype=oneflow.int64)
-
-    """
-
     axis_checked = _check_axis(dim, input.shape)
     if len(axis_checked) == 0:
         return input
@@ -65,35 +36,6 @@ def max_op(input, dim=None, keepdim=False):
 
 
 def min_op(input, dim=None, keepdim=False):
-    """Computes the minimum value of all elements in the input tensor.
-    
-    Args:
-        input (oneflow.Tensor): the Input Tensor
-        dim (int, optional): the dimension to reduce. Default: `None`
-        keepdim (bool, optional): whether the output tensor has dim retained or not. Default: `False`
-
-    Returns:
-        Tensor or Tuple(oneflow.Tensor, oneflow.Tensor(dtype=int64)): If :attr:`dim` is `None`, returns 
-        the minimum value of all elements in the `input` tensor. Otherwise, returns a tuple of Tensor (values, indices), 
-        where the `values` are the minimum value of all elements in the `input` tensor,
-        the `indices` are the indices of the elements in the original input tensor.
-    
-    For example:
-
-    .. code-block:: python
-
-        >>> import oneflow as flow
-        >>> input = flow.Tensor([[4, 1, 5], [2, 6, 3]])
-        >>> flow.min(input)
-        tensor(1., dtype=oneflow.float32)
-        >>> (values, indices) = flow.min(input, dim=1)
-        >>> values
-        tensor([1., 2.], dtype=oneflow.float32)
-        >>> indices
-        tensor([1, 0], dtype=oneflow.int64)
-
-    """
-
     axis_checked = _check_axis(dim, input.shape)
     if len(axis_checked) == 0:
         return input
@@ -107,23 +49,6 @@ def min_op(input, dim=None, keepdim=False):
 
 
 def sum_op(input, dim=None, keepdim=False):
-    """Computes the sum of row of elements in a tensor in the given axis, if the axis is None, sum of all elements will be caculated.
-    
-    For example:
-
-    .. code-block:: python
-
-        >>> import oneflow as flow
-        >>> input = flow.Tensor([[1, 2, 3], [4, 5, 6]])
-        >>> flow.sum(input)
-        tensor(21., dtype=oneflow.float32)
-        >>> flow.sum(input, dim=0)
-        tensor([5., 7., 9.], dtype=oneflow.float32)
-        >>> flow.sum(input, dim=1)
-        tensor([ 6., 15.], dtype=oneflow.float32)
-
-    """
-
     axis_checked = _check_axis(dim, input.shape)
     if len(axis_checked) == 0:
         return input
@@ -131,22 +56,6 @@ def sum_op(input, dim=None, keepdim=False):
 
 
 def mean_op(input, dim=None, keepdim=False):
-    """Computes the mean of row of elements in a tensor in the given axis, if the axis is None, mean of all elements will be caculated.
-    
-    For example:
-
-    .. code-block:: python
-
-        >>> import oneflow as flow
-        >>> input = flow.Tensor([[1, 2, 3], [4, 5, 6]])
-        >>> flow.mean(input)
-        tensor(3.5000, dtype=oneflow.float32)
-        >>> flow.mean(input, dim=0)
-        tensor([2.5000, 3.5000, 4.5000], dtype=oneflow.float32)
-        >>> flow.mean(input, dim=1)
-        tensor([2., 5.], dtype=oneflow.float32)
-
-    """
     axis_checked = _check_axis(dim, input.shape)
     if len(axis_checked) == 0:
         return input
@@ -154,28 +63,6 @@ def mean_op(input, dim=None, keepdim=False):
 
 
 def prod_op(input, dim=None, keepdim=False):
-    r"""Computes the product of row of elements in a tensor in the given axis.
-    
-    note: `if the dim is None, it will return a tensor with only one element whose value is the product of all elements of input.`
-
-    Args:
-        input (Tensor): the source tensor
-        dim (int): the axis along which to prod
-
-    For example:
-
-    .. code-block:: python
-
-        >>> import oneflow as flow
-        >>> input = flow.Tensor([[1, 2, 3], [4, 5, 6]])
-        >>> flow.prod(input)
-        tensor(720., dtype=oneflow.float32)
-        >>> flow.prod(input, dim=0)
-        tensor([ 4., 10., 18.], dtype=oneflow.float32)
-        >>> flow.prod(input, dim=1)
-        tensor([  6., 120.], dtype=oneflow.float32)
-
-    """
     axis_checked = _check_axis(dim, input.shape)
     if len(axis_checked) == 0:
         return input
